+++ packages/cli/api-importers/v3-importer-commons/src/AbstractSpecConverter.ts
@@ -83,6 +83,7 @@ export abstract class AbstractSpecConverter<
                 hasPaginatedEndpoints: false,
                 hasStreamingEndpoints: false,
                 isAuthMandatory: true,
+                idempotencyKeyGeneration: undefined,
                 platformHeaders: {
                     language: "",
                     sdkName: "",