+++ seed/java-sdk/unions-with-local-date/default/src/main/java/com/seed/unions/core/ClientOptions.java
@@ -47,7 +47,7 @@ public final class ClientOptions {
         this.headers.putAll(headers);
         this.headers.putAll(new HashMap<String, String>() {
             {
-                put("User-Agent", "com.fern:unions-with-local-date/0.0.1");
+                put("User-Agent", "com.fern.unions-with-local-date/0.0.1");
                 put("X-Fern-Language", "JAVA");
                 put("X-Fern-SDK-Name", "com.seed.fern:unions-sdk");
             }